+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/e2e/playwright/utils/utfyllingUtils.ts
@@ -71,8 +71,8 @@ const lastOppLegeerklæring = async (page: Page) => {
 
 const fyllUtMedlemsskap = async (page: Page) => {
     await page.getByRole('heading', { name: 'Medlemskap i folketrygden' });
-    await page.getByTestId('medlemskap-annetLandSiste12').getByText('Nei').click();
-    await page.getByTestId('medlemskap-annetLandNeste12_no').check();
+    await page.getByRole('group', { name: 'Har du bodd i utlandet i hele' }).getByLabel('Nei').check();
+    await page.getByRole('group', { name: 'Planlegger du å bo i utlandet' }).getByLabel('Nei').check();
     await page.getByTestId('typedFormikForm-submitButton').click();
 };