+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/søknad/SøknadRouter.tsx
@@ -18,7 +18,7 @@ import OppsummeringStep from './steps/oppsummering/OppsummeringStep';
 import KvitteringPage from '../pages/kvittering/KvitteringPage';
 import DineBarnStep from './steps/dine-barn/DineBarnStep';
 import DeltBostedStep from './steps/delt-bosted/DeltBostedStep';
-import { useVerifyUserOnWindowFocus } from '@navikt/sif-common-soknad-ds/src';
+import { useVerifyUserOnWindowFocus } from '@navikt/sif-common-soknad-ds';
 import søkerEndpoint from '../api/endpoints/søkerEndpoint';
 
 const SøknadRouter = () => {