+++ packages/sif-common-forms-ds/src/forms/medlemskap/stories/MedlemskapForm.stories.tsx
@@ -1,0 +1,18 @@
+import { Meta, StoryObj } from '@storybook/react';
+import { withIntl } from '../../../../storybook/decorators/withIntl';
+import MedlemskapFormExample from './MedlemskapFormExample';
+
+const meta: Meta<typeof MedlemskapFormExample> = {
+    component: MedlemskapFormExample,
+    title: 'Stegskjema/Medlemskap',
+    decorators: [withIntl],
+};
+
+export default meta;
+
+type Story = StoryObj<typeof MedlemskapFormExample>;
+
+export const Default: Story = {
+    name: 'Default',
+    render: () => <MedlemskapFormExample />,
+};