+++ packages/sif-common-soknad-ds/src/setup/SoknadApplication.tsx
@@ -32,6 +32,8 @@ interface Props {
     appKey: string;
     /** App name - not visual to user */
     appName: string;
+    /** App title - visaul to user*/
+    appTitle: string;
     /** Public path used in BrowserRouter */
     publicPath: string;
     /** Locale messages */
@@ -72,6 +74,7 @@ const SoknadApplication = ({
     useAmplitude,
     useLanguageSelector,
     children,
+    appTitle,
     onResetSoknad,
 }: Props) => {
     const [locale, setLocale] = React.useState<Locale>(localeFromSessionStorage);
@@ -90,7 +93,7 @@ const SoknadApplication = ({
                 telemetryCollectorURL={telemetryCollectorURL}
                 appVersion={appVersion}
                 isActive={useFaro}>
-                <ErrorBoundary appKey={appKey} onResetSoknad={onResetSoknad}>
+                <ErrorBoundary appKey={appKey} onResetSoknad={onResetSoknad} appTitle={appTitle}>
                     <AmplitudeProvider applicationKey={appKey} isActive={getUseAmplitude(useAmplitude)}>
                         <IntlProvider
                             locale={locale === 'nb' ? getBokmålLocale() : getNynorskLocale()}