+++ packages/appstatus-react-ds/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @navikt/appstatus-react-ds
 
+## 2.19.4
+
+### Patch Changes
+
+-   16f2ebf: Pakkeoppdateringer minor/patch
+    -   @navikt/sif-common-hooks@0.2.0
+
+## 2.19.3
+
+### Patch Changes
+
+-   485d6eb: Pakkeoppdateringer - minor/patch
+
+## 2.19.2
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+    -   @navikt/sif-common-hooks@0.2.0
+
 ## 2.19.1
 
 ### Patch Changes